+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/test/java/io/flutter/plugins/googlemaps/ConvertTest.java
@@ -46,7 +46,6 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.MockedStatic;
 import org.mockito.MockitoAnnotations;
 import org.robolectric.RobolectricTestRunner;
 
@@ -61,7 +60,6 @@ public class ConvertTest {
   @Mock private FlutterInjectorWrapper flutterInjectorWrapper;
 
   @Mock private GoogleMapOptionsSink optionsSink;
-  @Mock private MarkerOptionsSink markerOptionsSink;
 
   AutoCloseable mockCloseable;
 
@@ -127,28 +125,6 @@ public class ConvertTest {
     Assert.assertEquals(2, markerIds.size());
     Assert.assertEquals(marker1.markerId(), markerIds.get(0));
     Assert.assertEquals(marker2.markerId(), markerIds.get(1));
-  }
-
-  @Test
-  public void ConvertMarkerFromPigeonReturnsCorrectZIndex() {
-    try (MockedStatic<BitmapDescriptorFactory> mockedStatic =
-        mockStatic(BitmapDescriptorFactory.class)) {
-      mockedStatic.when(BitmapDescriptorFactory::defaultMarker).thenReturn(mockBitmapDescriptor);
-
-      double zIndex = 2.0;
-      long zIndexInt = 4;
-
-      float density = 1.0F;
-
-      Messages.PlatformMarker pMarker = createDefaultMarker();
-      pMarker.setZIndex(zIndex);
-      pMarker.setZIndexInt(zIndexInt);
-
-      Convert.interpretMarkerOptions(
-          pMarker, markerOptionsSink, assetManager, density, bitmapDescriptorFactoryWrapper);
-
-      verify(markerOptionsSink).setZIndex(zIndexInt);
-    }
   }
 
   @Test
@@ -810,38 +786,6 @@ public class ConvertTest {
         Convert.groundOverlayToPigeon(mockGroundOverlay, overlayId, true);
 
     assertGroundOverlayEquals(result, mockGroundOverlay, overlayId, null, bounds);
-  }
-
-  private Messages.PlatformMarker createDefaultMarker() {
-    Messages.PlatformDoublePair anchor =
-        new Messages.PlatformDoublePair.Builder().setX(0.0).setY(0.0).build();
-
-    Messages.PlatformBitmapDefaultMarker bitmapDefaultMarker =
-        new Messages.PlatformBitmapDefaultMarker();
-    Messages.PlatformBitmap icon =
-        new Messages.PlatformBitmap.Builder().setBitmap(bitmapDefaultMarker).build();
-
-    Messages.PlatformInfoWindow infoWindow =
-        new Messages.PlatformInfoWindow.Builder().setAnchor(anchor).build();
-
-    Messages.PlatformLatLng position =
-        new Messages.PlatformLatLng.Builder().setLatitude(0.0).setLongitude(0.0).build();
-
-    return new Messages.PlatformMarker.Builder()
-        .setAlpha(0.0)
-        .setAnchor(anchor)
-        .setConsumeTapEvents(false)
-        .setDraggable(false)
-        .setFlat(false)
-        .setIcon(icon)
-        .setInfoWindow(infoWindow)
-        .setPosition(position)
-        .setRotation(0.0)
-        .setVisible(false)
-        .setZIndex(0.0)
-        .setZIndexInt(0L)
-        .setMarkerId("default-marker")
-        .build();
   }
 }