+++ packages/google_maps_flutter/google_maps_flutter_android/example/lib/place_marker.dart
@@ -250,10 +250,10 @@ class PlaceMarkerBodyState extends State<PlaceMarkerBody> {
 
   Future<void> _changeZIndex(MarkerId markerId) async {
     final Marker marker = markers[markerId]!;
-    final int current = marker.zIndexInt;
+    final double current = marker.zIndex;
     setState(() {
       markers[markerId] = marker.copyWith(
-        zIndexIntParam: current == 12 ? 0 : current + 1,
+        zIndexParam: current == 12.0 ? 0.0 : current + 1.0,
       );
     });
   }