+++ packages/google_maps_flutter/google_maps_flutter_android/lib/src/google_maps_flutter_android.dart
@@ -793,7 +793,6 @@ class GoogleMapsFlutterAndroid extends GoogleMapsFlutterPlatform {
       rotation: marker.rotation,
       visible: marker.visible,
       zIndex: marker.zIndex,
-      zIndexInt: marker.zIndexInt,
       markerId: marker.markerId.value,
       clusterManagerId: marker.clusterManagerId?.value,
     );