+++ packages/google_maps_flutter/google_maps_flutter_android/lib/src/messages.g.dart
@@ -516,7 +516,6 @@ class PlatformMarker {
     this.rotation = 0.0,
     this.visible = true,
     this.zIndex = 0.0,
-    this.zIndexInt = 0,
     required this.markerId,
     this.clusterManagerId,
   });
@@ -543,8 +542,6 @@ class PlatformMarker {
 
   double zIndex;
 
-  int zIndexInt;
-
   String markerId;
 
   String? clusterManagerId;
@@ -562,7 +559,6 @@ class PlatformMarker {
       rotation,
       visible,
       zIndex,
-      zIndexInt,
       markerId,
       clusterManagerId,
     ];
@@ -582,9 +578,8 @@ class PlatformMarker {
       rotation: result[8]! as double,
       visible: result[9]! as bool,
       zIndex: result[10]! as double,
-      zIndexInt: result[11]! as int,
-      markerId: result[12]! as String,
-      clusterManagerId: result[13] as String?,
+      markerId: result[11]! as String,
+      clusterManagerId: result[12] as String?,
     );
   }
 }