+++ packages/google_maps_flutter/google_maps_flutter_android/pigeons/messages.dart
@@ -183,7 +183,6 @@ class PlatformMarker {
     this.rotation = 0.0,
     this.visible = true,
     this.zIndex = 0.0,
-    this.zIndexInt = 0,
     this.clusterManagerId,
   });
 
@@ -199,7 +198,6 @@ class PlatformMarker {
   final double rotation;
   final bool visible;
   final double zIndex;
-  final int zIndexInt;
   final String markerId;
   final String? clusterManagerId;
 }