+++ packages/google_maps_flutter/google_maps_flutter_web/example/integration_test/marker_test.dart
@@ -212,7 +212,6 @@ void main() {
       final gmaps.Marker newGmMarker = gmaps.Marker();
       const Marker marker = Marker(
         markerId: MarkerId('1'),
-        zIndex: 2.0,
         zIndexInt: 4,
       );