+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.9.2
+
+* Updates the example app to use the zIndexInt param instead of the deprecated zIndex.
+
 ## 2.9.1
 
 * Updates README to indicate that Andoid SDK <21 is no longer supported.