+++ packages/google_maps_flutter/google_maps_flutter_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.12 + 1
+
+* Deprecates `zIndex` parameter in Marker in favor of `zIndexInt`.
+
 ## 0.5.12
 
 * Adds support for ground overlay.